refactor: 사용하지 않는 메소드 제거 및 줄 정리
* refactor: CustomNumberPicker 필드 변수명 변경

* style: enum 클래스에 대한 패키지 구조 변경
