+++ src/main/java/racingcar/domain/car/condition/NameLength.java
@@ -1,4 +1,4 @@
-package racingcar.domain.car;
+package racingcar.domain.car.condition;
 
 public enum NameLength {
