+++ src/test/java/racingcar/service/picker/CustomNumberPicker.java
@@ -3,18 +3,17 @@ package racingcar.service.picker;
 import java.util.Iterator;
 import java.util.List;
 
-import racingcar.service.picker.NumberPicker;
-
 public class CustomNumberPicker implements NumberPicker {
 
-	private final Iterator<Integer> iterator;
+	private final Iterator<Integer> numbers;
 
 	public CustomNumberPicker(List<Integer> numbers) {
-		iterator = numbers.iterator();
+		this.numbers = numbers.iterator();
 	}
 
 	@Override
 	public int pickNumber() {
-		return iterator.next();
+		return numbers.next();
 	}
+
 }
