+++ src/test/java/racingcar/view/inputview/reader/CustomReader.java
@@ -4,14 +4,15 @@ import racingcar.view.input.reader.Reader;
 
 public class CustomReader implements Reader {
 
-	private String text = "";
+	private String text;
+
+	public void initText(String text) {
+		this.text = text;
+	}
 
 	@Override
 	public String readLine() {
 		return text;
 	}
 
-	public void setText(String text) {
-		this.text = text;
-	}
 }
